requirements.txt
# Core dependencies for Adaptive RAG
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-openai>=0.1.0  # For embeddings only (optional)
langchain-community>=0.0.1
langchain-huggingface>=0.1.0  # For HuggingFace embeddings
langchain-tavily>=0.3.0  # Updated Tavily integration
langgraph>=0.1.0
groq>=0.4.0
faiss-cpu>=1.7.4
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Embeddings (fallback when OpenAI not available)
sentence-transformers>=2.2.0
torch>=1.9.0

# Web search (legacy support)
tavily-python>=0.3.0

# Optional: For GPU support (uncomment if you have CUDA)
# faiss-gpu>=1.7.4
